requirements.txt
# Core dependencies
langchain>=0.1.1
langchain_community>=0.0.16
langchain-chroma>=0.1.0
sentence-transformers>=2.2.2
chromadb>=0.4.22
pydantic>=2.5.2

# LLM
ollama>=0.1.5

# Backend
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0

# Frontend
streamlit>=1.29.0

# Data processing
pandas>=2.1.3
numpy>=1.26.2

# Testing
pytest>=7.4.3

# Utilities
tqdm>=4.66.1